scraper_layer/requirements.txt
playwright>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.0
pillow>=9.5.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
pytest>=7.3.1
httpx>=0.24.1